pyproject.toml
[build-system]
requires = ["setuptools>=61", "packaging>=23"]
build-backend = "setuptools.build_meta"

[project]
name = "lab-automasion-ocr-module"
version = "0.1.0"
description = "Non-invasive Camera OCR wrapper for PaddleOCR"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Takaya Inoue" }]
maintainers = [{ name = "AUTO Kobo headquarters" }]
license = { text = "MIT" }
dependencies = [
  "customtkinter",
  "opencv-python",
  "Pillow",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/TITManagement/lab_automasion_OCR-module"
Repository = "https://github.com/TITManagement/lab_automasion_OCR-module"
Documentation = "https://github.com/TITManagement/lab_automasion_OCR-module#readme"

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.6", "mypy>=1.8"]

[project.scripts]
lab-ocr = "ocr_wrapper.run_ocr:main"
lab-camera-ocr = "ocr_wrapper.camera_ocr:main"
lab-camera-ocr-gui = "ocr_wrapper.camera_ocr_gui:main"
lab-ocr-generate-variants = "ocr_synthetic_data.generate_variants:main"
lab-ocr-generate-roi-strips = "ocr_dataset.source_cases.generate_roi_strips:main"
lab-ocr-prepare-source-case = "ocr_dataset.source_cases.prepare_source_case:main"
lab-ocr-source-case-gui = "ocr_dataset.source_cases.source_case_gui:main"

[tool.setuptools.packages.find]
where = ["ocr_wrapper/src", "ocr_synthetic_data/src", "ocr_dataset/src"]
include = ["ocr_wrapper*", "ocr_synthetic_data*", "ocr_dataset*"]

[tool.metadata]
copyright = "AUTO kobo"
